refactor(tui): split mod.rs (1012→32 lines) into event_loop, event_dispatch, submit
- event_loop.rs (142 lines): StartupResumeTarget + run_tui
- event_dispatch.rs (499 lines): dispatch_event
- submit.rs (244 lines): handle_submit + helpers
- mod.rs (32 lines): module declarations + re-exports

All files now under the 800-line limit. Follows the same flat
delegation pattern used by src/tui/runtime.rs.
